Refactor asset helpers into scoped modules
