+++ src/ui/cards/model/finance/index.js
@@ -2,7 +2,8 @@ import { getAssetState, getState } from '../../../../core/state.js';
 import { formatMoney } from '../../../../core/helpers.js';
 import { computeDailySummary } from '../../../../game/summary.js';
 import { getAssets, getHustles, getUpgrades } from '../../../../game/registryService.js';
-import { calculateAssetSalePrice, instanceLabel } from '../../../../game/assets/helpers.js';
+import { calculateAssetSalePrice } from '../../../../game/assets/actions.js';
+import { instanceLabel } from '../../../../game/assets/details.js';
 import { getKnowledgeProgress } from '../../../../game/requirements.js';
 import { buildSkillRewards, resolveTrack } from '../education.js';
 import { buildPulseSummary, computeTopEarner } from './pulse.js';
