+++ src/ui/cards/model/shopily.js
@@ -1,10 +1,8 @@
 import { ensureArray, formatHours, formatMoney } from '../../../core/helpers.js';
 import { getAssetState, getState } from '../../../core/state.js';
-import {
-  calculateAssetSalePrice,
-  instanceLabel,
-  formatMaintenanceSummary
-} from '../../../game/assets/helpers.js';
+import { calculateAssetSalePrice } from '../../../game/assets/actions.js';
+import { instanceLabel } from '../../../game/assets/details.js';
+import { formatMaintenanceSummary } from '../../../game/assets/maintenance.js';
 import {
   assignInstanceToNiche,
   getAssignableNicheSummaries,
