+++ src/ui/cards/model/videotube.js
@@ -1,6 +1,7 @@
 import { ensureArray, formatHours, formatMoney } from '../../../core/helpers.js';
 import { getAssetState, getState } from '../../../core/state.js';
-import { instanceLabel, formatMaintenanceSummary } from '../../../game/assets/helpers.js';
+import { instanceLabel } from '../../../game/assets/details.js';
+import { formatMaintenanceSummary } from '../../../game/assets/maintenance.js';
 import {
   assignInstanceToNiche,
   getAssignableNicheSummaries,
@@ -16,7 +17,7 @@ import {
   getQualityLevel,
   getQualityTracks
 } from '../../../game/assets/quality.js';
-import { setAssetInstanceName } from '../../../game/assets/helpers.js';
+import { setAssetInstanceName } from '../../../game/assets/actions.js';
 import { describeAssetLaunchAvailability } from './assets.js';
 import { registerModelBuilder } from '../modelBuilderRegistry.js';
